requirements.txt
alembic==1.4.2
Authlib==0.14.3
fastapi==0.61.0
celery==4.4.7
redis==3.5.3
httpx==0.14.0
ipython==7.17.0
itsdangerous==1.1.0
Jinja2==2.11.2
psycopg2==2.8.5
pytest==6.0.1
requests==2.24.0
SQLAlchemy==1.3.18
starlette==0.13.7
uvicorn==0.11.8
passlib==1.7.2
bcrypt==3.1.7
sqlalchemy-utils==0.36.8
python-multipart==0.0.5
pyjwt==1.7.1
numpy==1.19.2
aiofiles==0.5.0
Pillow==7.2.0